+++ Game/Game/Sourse/Actor/SlowArea.cpp
@@ -36,7 +36,7 @@ void SlowArea::Update()
 void SlowArea::Draw()
 {
 	Display::Instance()->SetScreen(PlayerBullet_Screen);
-	DrawCircle(position->x, position->y, size / 2, GetColor(255, 255, 255), 0);
+	DrawCircle((int)(position->x), (int)(position->y), size / 2, GetColor(255, 255, 255), 0);
 }
 
 void SlowArea::Hit(GameObject * obj)
