+++ Game/Game/Sourse/Actor/SlowBullet.cpp
@@ -44,7 +44,7 @@ void SlowBullet::Update()
 void SlowBullet::Draw()
 {
 	Display::Instance()->SetScreen(PlayerBullet_Screen);
-	DrawCircle(position->x, position->y, 5, GetColor(255, 255, 255), 0);
+	DrawCircle((int)(position->x), (int)(position->y), 5, GetColor(255, 255, 255), 0);
 }
 
 void SlowBullet::Hit(GameObject * obj)
